backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
pyjwt
fastapi
uvicorn
pydantic
gunicorn
